+++ TRAP/src/Graphics/API/ResourceLoader.h
@@ -66,13 +66,13 @@ namespace TRAP::Graphics::API
 		static RendererAPI::ResourceLoaderDesc DefaultResourceLoaderDesc;
 		static void StreamerThreadFunc(ResourceLoader* loader);
 		static uint64_t UtilGetTextureRowAlignment();
-		static uint32_t UtilGetTextureSubresourceAlignment(TRAP::Graphics::API::ImageFormat fmt = TRAP::Graphics::API::ImageFormat::Undefined);
-		static uint32_t UtilGetSurfaceSize(TRAP::Graphics::API::ImageFormat fmt, uint32_t width, uint32_t height,
-		                                   uint32_t depth, uint32_t rowStride, uint32_t sliceStride,
+		static uint64_t UtilGetTextureSubresourceAlignment(TRAP::Graphics::API::ImageFormat fmt = TRAP::Graphics::API::ImageFormat::Undefined);
+		static uint64_t UtilGetSurfaceSize(TRAP::Graphics::API::ImageFormat fmt, uint32_t width, uint32_t height,
+		                                   uint32_t depth, uint64_t rowStride, uint64_t sliceStride,
 										   uint32_t baseMipLevel, uint32_t mipLevels, uint32_t baseArrayLayer,
 										   uint32_t arrayLayers);
 		static bool UtilGetSurfaceInfo(uint32_t width, uint32_t height, TRAP::Graphics::API::ImageFormat fmt,
-		                               uint32_t* outNumBytes, uint32_t* outRowBytes, uint32_t* outNumRows);
+		                               uint64_t* outNumBytes, uint64_t* outRowBytes, uint64_t* outNumRows);
 		static RendererAPI::MappedMemoryRange AllocateUploadMemory(uint64_t memoryRequirement, uint32_t alignment);
 		void QueueBufferBarrier(const TRAP::Ref<Buffer>& buffer, RendererAPI::ResourceState state, SyncToken* token);
 		void QueueBufferUpdate(const RendererAPI::BufferUpdateDesc& desc, SyncToken* token);
